cbc first commit

+++ auth-service/src/main/java/com/jobspring/auth/dto/UserDTO.java
@@ -11,5 +11,5 @@ public class UserDTO {
     private Long id;
     private String email;
     private String fullName;
-    private Byte role;
+    private Integer role;
 }
